internal/utils: relocated the `character` and `string` namespaces from internal/interface into internal/utils.
Dropped the `character` and `string` namespaces from internal/interface, in
order to move them into internal/utils. The logic for this seems to make
more sense as internal/interface is more for things like types and classes
that help with different aspects of IDA and internal/utils consists of just
pure-python helpers.
